+++ core/common/common_config.js
@@ -14,7 +14,20 @@ const resolve = (_config_) => {
     config.isTS = fs.existsSync(path.resolve(projectPath, 'tsconfig.json'))
   }
 
-  config.webpack && typeof config.webpack.VueChunkStyle === 'undefined' && (config.webpack.VueChunkStyle = false)
+  // 增加默认值
+  typeof config['ES.Next'] === 'undefined' && (config['ES.Next'] = true)
+
+  if (config.webpack) {
+    typeof config.webpack.VueChunkStyle === 'undefined' && (config.webpack.VueChunkStyle = false)
+  }
+
+  if (config['workflow.dev']) {
+    typeof config['workflow.dev']['hot.reload'] === 'undefined' && (config['workflow.dev']['hot.reload'] = true)
+  }
+
+  if (config['workflow.build']) {
+    typeof config['workflow.build']['cache'] === 'undefined' && (config['workflow.build']['cache'] = 'hash')
+  }
 
   const { cache } = config[ 'workflow.build' ]